requirements.txt
aiohttp==3.7.4.post0
antiorm==1.2.1
APScheduler==3.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
boto3==1.24.43
botocore==1.27.43
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clarifai==2.6.2
configparser==3.8.1
cryptography==3.4.7
cycler==0.11.0
db==0.1.1
db.py==0.5.3
dill==0.3.4
discord.py==1.6.0
distlib==0.3.4
docopt==0.6.2
EasyProcess==0.3
emoji==1.2.0
eval==0.0.5
filelock==3.4.2
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
googleapis-common-protos==1.53.0
grpcio==1.37.1
h11==0.13.0
huggingface-hub==0.8.1
idna==2.10
ignite==1.1.0
importlib-metadata==4.12.0
imutils==0.5.4
instapy==0.6.13
jmespath==1.0.1
jsonschema==2.6.0
kiwisolver==1.3.2
MeaningCloud-python==2.0.0
multidict==5.1.0
numpy==1.22.0
onnx==1.11.0
onnxruntime==1.11.1
opencv-python==4.5.3.56
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pbr==5.8.0
Pillow==9.0.0
platformdirs==2.4.1
plyer==2.0.0
prettytable==0.7.2
protobuf==3.17.0
psycopg2-binary==2.9.3
pybars3==0.9.7
pycparser==2.20
PyMeta3==0.5.1
pyOpenSSL==20.0.1
pyparsing==3.0.6
PySocks==1.7.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-telegram-bot==13.5
pytorch-pretrained-bert==0.6.2
pytz==2021.1
PyVirtualDisplay==2.1
PyWavelets==1.2.0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scipy==1.7.3
selenium==4.3.0
sentry-sdk==1.7.2
seq2seq==0.1.6
SimpleITK==2.1.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
stevedore==3.5.0
tokenizers==0.12.1
torch==1.10.1
torchtext==0.11.1
tornado==6.1
tqdm==4.60.0
train==0.0.5
transformers==4.21.0
trio==0.21.0
trio-websocket==0.9.2
tweet-preprocessor==0.6.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webdriverdownloader==1.1.0.3
wrapt==1.13.3
wsproto==1.1.0
yarl==1.6.3
zipp==3.8.1